codebase-token-counter/requirements.txt
gitpython==3.1.41
tqdm==4.66.3
transformers==4.38.1
rich==13.9.4

# Test dependencies
pytest==8.0.0
pytest-cov==4.1.0
